resize function ok

+++ src/UserVideoAbstraction.cpp
@@ -1,7 +1,7 @@
 #include "VideoAbstraction.h"
 #include "UserVideoAbstraction.h"
 
-UserVideoAbstraction::UserVideoAbstraction(const char* inputpath, const char* out_path, const char* log_path, const char* config_path,const char* index_path,const char* videoname, const char* midname, int scale)
+UserVideoAbstraction::UserVideoAbstraction(const char* inputpath, const char* out_path, const char* log_path, const char* config_path,const char* index_path,const char* videoname, const char* midname, float scale)
 {
 	string t1=inputpath;
 	string t2=out_path;
